+++ R/class-args.R
@@ -49,6 +49,11 @@
 #' @field resolved `[logical]`\cr Has the object been resolved
 #' @field value `[ANY]`\cr The resolve value
 #' @field stop `[character]`\cr `"none"`, `"hard"`, or `"soft"`
+#' @field execute `[function]`\cr (For advanced use).  A `function` to be
+#'   evaluated along with the arg.  The function can have no parameters, a
+#'   single parameter for the [scribeArg] object, or accept the [scribeArg]
+#'   object as its first argument, and the [scribeCommandArgs] object as its
+#'   second.  Both objects will be passed by position
 #'
 #' @examples
 #' # new_arg() is recommended over direct use of scribeArg$new()
@@ -80,7 +85,8 @@ scribeArg <- methods::setRefClass( # nolint: object_name_linter.
     positional = "logical",
     resolved   = "logical",
     value      = "ANY",
-    stop       = "character"
+    stop       = "character",
+    execute    = "function"
   )
 )
 
@@ -93,7 +99,8 @@ scribeArg$methods(
     n       = NA_integer_,
     info    = NA_character_,
     options = list(),
-    stop    = c("none", "hard", "soft")
+    stop    = c("none", "hard", "soft"),
+    execute = invisible
   ) {
     "
     Initialize the \\link{scribeArg} object
@@ -101,7 +108,7 @@ scribeArg$methods(
     See \\strong{fields} for parameter information.
     "
     arg_initialize(
-      .self,
+      self    = .self,
       aliases = aliases,
       action  = action,
       default = default,
@@ -109,7 +116,8 @@ scribeArg$methods(
       n       = n,
       info    = info,
       options = options,
-      stop    = stop
+      stop    = stop,
+      execute = execute
     )
   },