+++ R/command-args.R
@@ -20,7 +20,7 @@
 #' @export
 command_args <- function(
     x = NULL,
-    include = c("help", "version", NA_character_),
+    include = getOption("scribe.include", c("help", "version", NA_character_)),
     string = NULL
 ) {
   if (is.null(string)) {
@@ -157,6 +157,7 @@ ca_resolve <- function(self) {
   on.exit(
     expr =  if (!self$resolved) {
       self$field("working", self$get_input())
+      self$field("stop", "none")
     },
     add = TRUE
   )
@@ -207,42 +208,20 @@ ca_resolve <- function(self) {
 
   self$field("values", self$values[order(arg_order)])
   self$field("resolved", TRUE)
-
-  if ("help" %in% self$included) {
-    m <- match("help", names(self$values), 0L)
-    if (isTRUE(self$values$help)) {
-      self$help()
-      exit()
-      return(invisible(self))
-    }
-    self$field("values", self$values[-m])
-  }
-
-  if ("version" %in% self$included) {
-    m <- match("version", names(self$values), 0L)
-    if (isTRUE(self$values$version)) {
-      self$version()
-      exit()
-      return(invisible(self))
-    }
-    self$field("values", self$values[-m])
-  }
-
   invisible(self)
 }
 
 ca_parse <- function(self) {
   self$resolve()
-  values <- self$get_values()
 
-  # clean up names
-  regmatches(names(values), regexpr("^-+", names(values))) <- ""
-  regmatches(names(values), gregexpr("-", names(values))) <- "_"
-
-  if (any(names(values)[vapply(values, isTRUE, NA)] %in% self$included)) {
-    return(invisible(values))
+  for (arg in self$get_args()) {
+    ca_do_execute(self, arg)
   }
 
+  # clean up names
+  values <- self$get_values()
+  regmatches(names(values), regexpr("^-+", names(values))) <- ""
+  regmatches(names(values), gregexpr("-", names(values))) <- "_"
   values
 }
 
@@ -284,14 +263,15 @@ ca_get_args <- function(self, included = TRUE) {
 
 ca_add_argument <- function(
     self,
-    ...,
-    n = NA_integer_,
-    action = NULL,
-    convert = default_convert,
-    options = NULL,
-    default = NULL,
-    info = NULL,
-    stop = "none"
+  ...,
+  n = NA_integer_,
+  action = NULL,
+  convert = default_convert,
+  options = NULL,
+  default = NULL,
+  info = NULL,
+  stop = "none",
+  execute = invisible
 ) {
   if (is_arg(..1)) {
     arg <- ..1
@@ -317,7 +297,8 @@ ca_add_argument <- function(
       default = default,
       info = info,
       stop = stop,
-      n = as.integer(n)
+      n = as.integer(n),
+      execute = execute
     )
   }
 
@@ -371,6 +352,21 @@ ca_add_example <- function(self, x = NULL, comment = "", prefix = "$ ") {
 
 ca_get_examples <- function(self) {
   self$examples
+}
+
+ca_do_execute <- function(self, arg) {
+  n <- length(formals(arg$execute))
+
+  # switch(int) doesn't account for defaults
+  if (n == 0L) {
+    arg$execute()
+  } else if (n == 1L) {
+    arg$execute(arg)
+  } else {
+    arg$execute(arg, self)
+  }
+
+  invisible(self)
 }
 
 # internal ----------------------------------------------------------------